Setup Mailchimp API
This sets up the Python API for MailChimp. It adds its dependencies to
requirements.txt

It also implements a simple test to print out all subscribers from a
given list. It uses a list ID, user name, and API key stored in a conf
file, which should never be commited.