Refactor to use MailChimp API more effectively
This PR changes the model of how the subscribtion status of clients are
checked.

First, it introduces a Client class to encapsulate data about clients.

Second, it restructures the check to match the business logic of event
participant registration.  It makes individual calls out on the API for
each client. If a client is pending or not present on the list, they are
sent an opt-in message from MailChimp. For all other states, no action
is taken.