Refactor Client class, write to file, and property based testing
This commit primarily adds the functionality to write out the clients to
a CSV file. This functionality is now the default. If a user wishes to
have Mailchimp send a file, you'll have to set it as a config option.

This commit also refactors the Client class constructor to reduce the
coupling bewteen Client and the file row object. It now takes in the
params one value at a time, rather than reading a whole row. The row
could be malformed, and it doesn't make much sense to have the class
validate this.

Lastly, this commit introduces property based testing for the new write
to file functionality. It uses a combination of Hypothesis to generate
randomized Client objects and mocking to capture calls to the OS.